feat: implement native permission handling in Kiro and Grok drivers
